[#154] ref: 지원님 피드백 반영 (프로토콜 분리)
