+++ ThingLog/ViewModel/UserInformationViewModel/UserInformationiCloudViewModel.swift
@@ -6,21 +6,6 @@
 //
 import Foundation
 
-/// 유저정보의 데이터와 업데이트를 하기 위한 메소드들을 정의한 프로토콜이다.
-protocol UserInformationViewModelable {
-    /// 유저정보를 가져오도록 한다.
-    func fetchUserInformation(completion: @escaping (UserInformationable?) -> Void)
-    
-    /// 유저정보를 업데이트 하도록 한다.
-    func updateUserInformation(_ user: UserInformationable)
-    
-    /// 테스트용으로 유저정보를 초기화하도록 한다.
-    func resetUserInformation()
-    
-    /// 유저정보가 변경된 경우 감지하여 추가적인 로직을  completion에서 구현하도록 한다.
-    func subscribeUserInformationChange(completion: @escaping (UserInformationable?) -> Void)
-}
-
 /// `KeyValueStorage`를 이용하여 유저정보를 가져오고 업데이트하는 ViewModel
 final class UserInformationiCloudViewModel: UserInformationViewModelable {
     func fetchUserInformation(completion: @escaping (UserInformationable?) -> Void) {
